Moved most systems and states from localspacedemo to base space trader app
